try fixing env failure on CI
